Just fixed a few lines of code which had the wrong datatypes

+++ employeeTwo.c
@@ -53,7 +53,7 @@ PtrToEmployee searchEmployeeByPhone(PtrToConstEmployee ptr, int size, char* phon
     return searchEmployeeTable(ptr, size, phone, compareEmployeePhone);
 }
 
-PtrToEmployee serchEmployeebyNumber(PtrToConstEmployee ptr, int size, long salary)
+PtrToEmployee serchEmployeebyNumber(PtrToConstEmployee ptr, int size, double salary)
 {
     return searchEmployeeTable(ptr, size, &salary, compareEmployeeNumber);
 }
